requirements.txt
orjson==3.4.6
alembic==1.4.3
asyncpg==0.21.0
fastapi==0.62.0
gino==1.0.1
gino-starlette==0.1.1
psycopg2-binary==2.8.6
uvicorn==0.13.2
pycryptodomex==3.9.9
